requirements.txt
groq==0.4.1
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.11
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pynput==1.7.6
psutil==5.9.6
python-dotenv==1.0.0
pyperclip==1.8.2

# Vision & Image Processing
mss==9.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1

# UI & GUI
PyQt5==5.15.9
PyQt5-sip==12.13.0

# Additional utilities
scipy==1.11.3
